+++ src/effects/spells/holyAura.js
@@ -1,4 +1,4 @@
-import { baseSpellEffect, generateATLChange, spellEffectModules } from "../specialSpells.js";
+import { baseSpellEffect, generateATLChange } from "../specialSpells.js";
 
 export function holyAuraEffect(document) {
   let effect = baseSpellEffect(document, document.name);
@@ -17,7 +17,7 @@ export function holyAuraEffect(document) {
     },
   );
 
-  if (spellEffectModules().atlInstalled) {
+  if (CONFIG.DDBI.EFFECT_CONFIG.SPELLS.installedModules.atlInstalled) {
     effect.changes.push(generateATLChange("ATL.dimLight", CONST.ACTIVE_EFFECT_MODES.UPGRADE, '5'));
     effect.changes.push(generateATLChange("ATL.lightColor", CONST.ACTIVE_EFFECT_MODES.OVERRIDE, '#ffffff'));
     effect.changes.push(generateATLChange("ATL.lightAlpha", CONST.ACTIVE_EFFECT_MODES.OVERRIDE, '0.25'));