+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 3.1.7
 
 * Spell Effect: Aid - could add more HP than it should in some combinations.
+* Spell Effect fixes: Shillelagh, Magic Weapon.
+* Feature Effect: Channel Divinity: Sacred Weapon.
 
 # 3.1.16